Add files via upload
The ingredients-cleaner.py program goes through the entire JSON database (allrecipes-recipes.json) and cleans the ingredients list, similar to the recipe-analyzer program, but instead of doing so for ingredients that go with a user-input ingredient, this program cleans the whole database. The cleaned ingredients list is added as a new key or item to each JSON line.